LPM: allow an LPAR to coordinate the LPM test
Currently the LPM tests are self-validating; the LPAR being migrated is also the LPAR running the test. This fix allows an external LPAR to test migration of another LPAR.

Signed-off-by: Matthew Merck <merckmatthew@gmail.com>

Create new yaml file to not affect previous configuration

added new YAML files and edited lpm.yaml to keep previous behavior with option to add "lpar" and "lpar_ip" values for peer testing

Signed-off-by: Matthew Merck <merckmatthew@gmail.com>

Start RSCT and RMC services in peer LPAR

Currently, the RSCT and RMC services are started in the host LPAR orchestrating the tests. They should instead be started on the peer LPAR

Signed-off-by: Matthew Merck <merckmatthew@gmail.com>